requirements.txt
aniso8601==9.0.1
click==8.0.1
colorama==0.4.4
docker-push-ssh==0.1.5
Flask==2.0.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
paho-mqtt==1.5.1
pandas==1.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
